Moved question library components out of professor folder
